+++ src/libtrx/game/console/cmd/play_level.c
@@ -4,31 +4,22 @@
 #include "game/game_string.h"
 #include "strings.h"
 
+static const GF_LEVEL *M_FindLevel(const char *user_input);
 static COMMAND_RESULT M_Entrypoint(const COMMAND_CONTEXT *ctx);
 
-static COMMAND_RESULT M_Entrypoint(const COMMAND_CONTEXT *const ctx)
+static const GF_LEVEL *M_FindLevel(const char *const user_input)
 {
-    if (String_IsEmpty(ctx->args)) {
-        return CR_BAD_INVOCATION;
+    int32_t level_num = -1;
+    if (String_ParseInteger(user_input, &level_num)) {
+        return GF_GetLevelByOrdinalNumber(GFLT_MAIN, level_num);
     }
 
-    VECTOR *source = nullptr;
-    VECTOR *matches = nullptr;
-    int32_t level_to_load = -1;
-    const GF_LEVEL_TABLE *const level_table = GF_GetLevelTable(GFLT_MAIN);
-
-    if (String_ParseInteger(ctx->args, &level_to_load)) {
-        if (GF_GetGymLevel() == nullptr) {
-            level_to_load -= 1;
-        }
-        goto matched;
-    }
-
-    source = Vector_Create(sizeof(STRING_FUZZY_SOURCE));
-    for (int32_t i = 0; i < level_table->count; i++) {
+    VECTOR *source = Vector_Create(sizeof(STRING_FUZZY_SOURCE));
+    for (int32_t i = 0; i < GF_GetLevelCount(GFLT_MAIN); i++) {
+        const GF_LEVEL *const level = GF_GetLevel(GFLT_MAIN, i);
         STRING_FUZZY_SOURCE source_item = {
-            .key = level_table->levels[i].title,
-            .value = (void *)(intptr_t)i,
+            .key = level->title,
+            .value = (void *)level,
             .weight = 1,
         };
         if (source_item.key != nullptr) {
@@ -40,47 +31,19 @@ static COMMAND_RESULT M_Entrypoint(const COMMAND_CONTEXT *const ctx)
     if (gym_level != nullptr) {
         STRING_FUZZY_SOURCE source_item = {
             .key = "gym",
-            .value = (void *)(intptr_t)gym_level->num,
+            .value = (void *)gym_level,
             .weight = 1,
         };
         Vector_Add(source, &source_item);
     }
 
-    COMMAND_RESULT result;
-    matches = String_FuzzyMatch(ctx->args, source);
-
-    if (matches->count == 0) {
-        Console_Log(GS(OSD_INVALID_LEVEL));
-        result = CR_FAILURE;
-        goto cleanup;
-    } else if (matches->count >= 1) {
+    const GF_LEVEL *result = nullptr;
+    VECTOR *matches = String_FuzzyMatch(user_input, source);
+    if (matches->count >= 1) {
         const STRING_FUZZY_MATCH *const match = Vector_Get(matches, 0);
-        level_to_load = (int32_t)(intptr_t)match->value;
-        goto matched;
+        result = (const GF_LEVEL *)match->value;
     }
 
-matched:
-    if (level_to_load >= 0 && level_to_load < level_table->count) {
-        const GF_LEVEL *const level = &level_table->levels[level_to_load];
-#if TR_VERSION == 1
-        if (level->type == GFL_DUMMY || level->type == GFL_CURRENT) {
-            Console_Log(GS(OSD_INVALID_LEVEL));
-            result = CR_FAILURE;
-            goto cleanup;
-        }
-#endif
-        GF_OverrideCommand((GF_COMMAND) {
-            .action = GF_SELECT_GAME,
-            .param = level_to_load,
-        });
-        Console_Log(GS(OSD_PLAY_LEVEL), level->title);
-        result = CR_SUCCESS;
-    } else {
-        Console_Log(GS(OSD_INVALID_LEVEL));
-        result = CR_FAILURE;
-    }
-
-cleanup:
     if (matches != nullptr) {
         Vector_Free(matches);
         matches = nullptr;
@@ -89,8 +52,30 @@ cleanup:
         Vector_Free(source);
         source = nullptr;
     }
-
     return result;
+}
+
+static COMMAND_RESULT M_Entrypoint(const COMMAND_CONTEXT *const ctx)
+{
+    if (String_IsEmpty(ctx->args)) {
+        return CR_BAD_INVOCATION;
+    }
+
+    VECTOR *matches = nullptr;
+    const GF_LEVEL *const level = M_FindLevel(ctx->args);
+
+    if (level == nullptr || level->type == GFL_DUMMY
+        || level->type == GFL_CURRENT) {
+        Console_Log(GS(OSD_INVALID_LEVEL));
+        return CR_FAILURE;
+    }
+
+    GF_OverrideCommand((GF_COMMAND) {
+        .action = GF_SELECT_GAME,
+        .param = level->num,
+    });
+    Console_Log(GS(OSD_PLAY_LEVEL), level->title);
+    return CR_SUCCESS;
 }
 
 REGISTER_CONSOLE_COMMAND("play", M_Entrypoint, GS_ID(CONSOLE_HELP_PLAY_LEVEL))